refactor(phase4): extract field_parsing from eon_parser.rs
Move `parse_field_type`, `parse_relation_document`, `parse_field_validation_document`,
`parse_field_transforms_document`, and related helpers (~301 lines) into
`eon_parser/field_parsing.rs`. Also includes `ParsedFieldType` struct and
`impl ScalarType { rust_type }` which live alongside the field-type parsing logic.

Follow-up: prune stale imports from `eon_parser.rs`, `policy_parsing.rs`,
`codegen.rs`, and `policy_plan_tokens.rs` left behind by earlier extractions.
All 233 passing tests remain green; 19 pre-existing failures unchanged.
